==> 01_predict_question_tags_stackoverflow/notebooks/01_preprocessing.ipynb <==
TRAIN_PATH = "../data/raw/train.tsv"
VAL_PATH = "../data/raw/validation.tsv"
TEST_PATH = "../data/raw/test.tsv"

TRAIN_EXPORT_PATH = "../data/processed/00_train_df.pkl"
VAL_EXPORT_PATH = "../data/processed/00_validation_df.pkl"
TEST_EXPORT_PATH = "../data/processed/00_test_df.pkl"

# load packages
import pandas as pd
import numpy as np

from ast import literal_eval
import re

import nltk
from nltk.corpus import stopwords

pd.set_option('display.max_colwidth', None)

___
## Functions:

def read_data(file_path):
    """
    Read tap separated data; we applied the literal_eval function
    on it so as to evaluate the exact data type of tags from the file
    """
    data = pd.read_csv(file_path, sep='\t')
    data['tags'] = data['tags'].apply(literal_eval)
    return data

REPLACE_BY_SPACE_RE = re.compile('[/(){}\[\]\|@,;]')
BAD_SYMBOLS_RE = re.compile('[^0-9a-z #+_]')
STOPWORDS = set(stopwords.words("english"))

def text_prepare(text):
    """
    text: a string to process
    return: modified initial string
    """
    text = text.lower()
    text = REPLACE_BY_SPACE_RE.sub(' ', text)
    text = BAD_SYMBOLS_RE.sub('', text)
    text = ' '.join([x for x in text.split() if x and x not in STOPWORDS])
    return text

def text_stemmer(text):
    """
    A stemmer functoin that takes a text as input; tokenize and then normalize it
    returns the same text after stemming
    """
    stemmer = nltk.stem.PorterStemmer()
    stems = [stemmer.stem(token) for token in text.split()]
    return " ".join(stems)

def text_lemmatizer(text):
    """
    A stemmer functoin that takes a text as input; tokenize and then normalize it
    returns the same text after Lemmatizing
    """
    lemmatizer = nltk.stem.WordNetLemmatizer()
    lemmas = [lemmatizer.lemmatize(token) for token in text.split()]
    return " ".join(lemmas)

___
## Read data

raw_train = read_data(TRAIN_PATH)
raw_val = read_data(VAL_PATH)
raw_test = pd.read_csv(TEST_PATH, sep='\t')

raw_train.sample(5)

raw_test.sample(5)

___
## Prepare title column:
One of the most known difficulities when working with natural data is that it's unstructured. For example, if used "as is" and extract tokens just by splitting the titles by whitespaces, there would be many weird tokens like (3.5?,"flip), etc. To prevent the problems, it's usually useful to prepare the data somehow. We are gonna use a RegEx to replace these weird tokens with whitespaces.

sample = np.random.randint(0, len(raw_train)+1)
print("Before:")
print(raw_train['title'].values[sample])
print()
print("After:")
print(text_prepare(raw_train['title'].values[sample]))

**Now preprocess title columns to acceptable form**

train_df = raw_train.copy()
val_df = raw_val.copy()
test_df = raw_test.copy()

train_df['title'] = train_df['title'].apply(text_prepare)
val_df['title'] = val_df['title'].apply(text_prepare)
test_df['title'] = test_df['title'].apply(text_prepare)

**Now a quick sanity check**

sample = np.random.randint(0, len(raw_train)+1)
pd.DataFrame([raw_train['title'][sample],
          train_df['title'][sample]],
         index=["raw", "processed"], columns=[f'sample ({sample})'])

## Export data

train_df.to_pickle(TRAIN_EXPORT_PATH)
val_df.to_pickle(VAL_EXPORT_PATH)
test_df.to_pickle(TEST_EXPORT_PATH)

___
## In case we want to normalize our text using stemming or lemmatizing:

train_df['title'][:20].apply(text_stemmer)

train_df['title'][:20].apply(text_lemmatizer)
