==> 01_predict_question_tags_stackoverflow/notebooks/02_feature_engineering.ipynb <==
TRAIN_PATH = "../data/processed/00_train_df.pkl"
VAL_PATH = "../data/processed/00_validation_df.pkl"
TEST_PATH = "../data/processed/00_test_df.pkl"

TRAIN_FEATURES_BOW = "../data/processed/01_train_features_BOW.pkl"
VAL_FEATURES_BOW = "../data/processed/01_validation_features_BOW.pkl"
TEST_FEATURES_BOW = "../data/processed/01_test_features_BOW.pkl"

TRAIN_FEATURES_TFIDF = "../data/processed/01_train_features_TFIDF.pkl"
VAL_FEATURES_TFIDF = "../data/processed/01_validation_features_TFIDF.pkl"
TEST_FEATURES_TFIDF = "../data/processed/01_test_features_TFIDF.pkl"


TRAIN_TARGET_EXPORT = "../data/processed/01_train_target.pkl"
VAL_TARGET_EXPORT = "../data/processed/01_validation_target.pkl"

# Load packages
import pandas as pd
import numpy as np

import pickle

from collections import Counter

from sklearn.feature_extraction.text import CountVectorizer
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.preprocessing import MultiLabelBinarizer

from scipy import sparse

pd.set_option('display.max_colwidth', None)
pd.options.display.max_columns = 1000

___
## Read Data:

train_df = pd.read_pickle(TRAIN_PATH)
val_df = pd.read_pickle(VAL_PATH)
test_df = pd.read_pickle(TEST_PATH)

train_df.head()

___
## First step: count tag/word frequency:
**At the first step we will get the count of each word and tag so we can sort them and find the best vectors we can classify upon**

tags_counts = Counter() # counter for tags
words_counts = Counter() # counter for words in titles

for tags in train_df['tags']:
    for tag in tags:
        tags_counts[tag] += 1

for words in train_df['title']:
    for word in words.split():
        words_counts[word] += 1

---
## Construct the BOW:

DICT_SIZE = 5000

# Find the N sorted words as my features
#SORT_BY_FREQ = sorted(words_counts.keys(), key=lambda x: words_counts[x], reverse=True)[:DICT_SIZE]
SORT_BY_FREQ = [x[0] for x in words_counts.most_common(DICT_SIZE)]

BOW_vectorizer = CountVectorizer(vocabulary=SORT_BY_FREQ)

X_train_BOW = BOW_vectorizer.fit_transform(train_df['title'])
X_val_BOW = BOW_vectorizer.transform(val_df['title'])
X_test_BOW = BOW_vectorizer.transform(test_df['title'])

print(X_train_BOW[5])
print(vectorizer.get_feature_names_out()[:10])

print("X_train_BOW.shape: ", X_train_BOW.shape)
print("X_val_BOW.shape: ", X_val_BOW.shape)
print("X_test_BOW.shape: ", X_test_BOW.shape)

___
## Construct TF-IDF:

tfidf_vectorizer = TfidfVectorizer(min_df=5, max_df=0.9,
                                   ngram_range=(1, 2),
                                   token_pattern='(\S+)')

X_train_tfidf = tfidf_vectorizer.fit_transform(train_df['title'])
X_val_tfidf = tfidf_vectorizer.transform(val_df['title'])
X_test_tfidf = tfidf_vectorizer.transform(test_df['title'])

print(X_train_tfidf[5])
print(X_train_tfidf.shape)
print(tfidf_vectorizer.get_feature_names_out()[:10])

# tfidf_reversed_vocab = {i:word for word, i in tfidf_vectorizer.vocabulary_.items()}

___
## MultiLabelBinarizer:
**Construct a matrix of targets using the multilabelbinarizer**

mlb = MultiLabelBinarizer(classes=sorted(tags_counts.keys()))

y_train = mlb.fit_transform(train_df['tags'])
y_val = mlb.transform(val_df['tags'])

pd.DataFrame(y_train, columns=mlb.classes_, index=train_df['tags'].index).head()

___
## Export and save results to use in modeling:

# Save BOW features
with open(TRAIN_FEATURES_BOW, 'wb') as handle:
    pickle.dump(X_train_BOW, handle)

with open(VAL_FEATURES_BOW, 'wb') as handle:
    pickle.dump(X_val_BOW, handle)

with open(TEST_FEATURES_BOW, 'wb') as handle:
    pickle.dump(X_test_BOW, handle)

# Save TF-IDF features
with open(TRAIN_FEATURES_TFIDF, 'wb') as handle:
    pickle.dump(X_train_tfidf, handle)

with open(VAL_FEATURES_TFIDF, 'wb') as handle:
    pickle.dump(X_val_tfidf, handle)

with open(TEST_FEATURES_TFIDF, 'wb') as handle:
    pickle.dump(X_test_tfidf, handle)

# Save MultiLabelBinarizer target
with open(TRAIN_TARGET_EXPORT, 'wb') as handle:
    pickle.dump(y_train, handle)

with open(VAL_TARGET_EXPORT, 'wb') as handle:
    pickle.dump(y_val, handle)

# Save the MLB
with open('../data/objects/mlb.pkl', 'wb') as handle:
    pickle.dump(mlb, handle)
    
# Save the BOW_vectorizer
with open('../data/objects/BOW_vectorizer.pkl', 'wb') as handle:
    pickle.dump(BOW_vectorizer, handle)
    
# Save the tfidf_vectorizer
with open('../data/objects/tfidf_vectorizer.pkl', 'wb') as handle:
    pickle.dump(tfidf_vectorizer, handle)
