==> 01_predict_question_tags_stackoverflow/notebooks/03_base_model.ipynb <==
TRAIN_PATH = "../data/processed/00_train_df.pkl"
VAL_PATH = "../data/processed/00_validation_df.pkl"
TEST_PATH = "../data/processed/00_test_df.pkl"

TRAIN_FEATURES_BOW = "../data/processed/01_train_features_BOW.pkl"
VAL_FEATURES_BOW = "../data/processed/01_validation_features_BOW.pkl"
TEST_FEATURES_BOW = "../data/processed/01_test_features_BOW.pkl"

TRAIN_FEATURES_TFIDF = "../data/processed/01_train_features_TFIDF.pkl"
VAL_FEATURES_TFIDF = "../data/processed/01_validation_features_TFIDF.pkl"
TEST_FEATURES_TFIDF = "../data/processed/01_test_features_TFIDF.pkl"


TRAIN_TARGET_EXPORT = "../data/processed/01_train_target.pkl"
VAL_TARGET_EXPORT = "../data/processed/01_validation_target.pkl"

# Load packages
import pandas as pd
import numpy as np

import pickle

from sklearn.pipeline import make_pipeline
from sklearn.multiclass import OneVsRestClassifier
from sklearn.linear_model import LogisticRegression, RidgeClassifier

from sklearn.metrics import accuracy_score, precision_score, recall_score, f1_score

___
## Functions:

def calculate_quality(ground_truth, predictions, metric_function, model_name, Average='macro'):
    """
    Calculate the quality of the model according to different metric scores
    Input:
        ground_truth: from real observed data
        predictions: the predicted values from the model
        metric_function: the metric score funcrion used to measure performance
    Output:
        A dict of all scores for the given inputs
    """
    quality_score = {}
    if metric_function.__name__ == 'accuracy_score':
        quality_score[model_name] = round(metric_function(ground_truth, predictions), 3)
    else:
        quality_score[model_name] = round(metric_function(ground_truth, predictions, average=Average), 3)

    quality_score = pd.Series(quality_score.values(), index=quality_score.keys())
    
    return quality_score

def evaluate_model(model, data_list, metrics_list, model_name, score_average='macro'):
    """
    Get the scores of the model to better understand its performance
    Input:
        data_list: a list of all data that we evaluate model upon, train and test data.
                typical input : [[X_train, y_train, 'train'], [X_test, y_test, 'test']]
        scores_list: a list of all metrics used in the evaluation. 
                typical input : [accuracy_score, precision_score, recall_score, f1_score]
        model_name: a string input used as the index for score dataframe.
    Output:
        scores: a dataframe of evaluation based on data.
        general_error: the generalized error that would be used for logging in mlflow
    """
    predicts = []
    scores = []
    for [X,y,stage] in data_list:

        probas = model.predict(X)
        predictions = pd.DataFrame(probas)
        predicts.append(predictions)

        result = {score.__name__:calculate_quality(y, predictions, score, f"{model_name}_{stage}", score_average)
                        for score in metrics_list}

        result = pd.concat(result, axis=1)
        scores.append(result)
    scores = pd.concat(scores)
    return scores, predicts

def print_words_for_tag(classifier, tag, tags_classes, index_to_words):
    """
        classifier: trained classifier
        tag: particular tag
        tags_classes: a list of classes names from MultiLabelBinarizer
        index_to_words: index_to_words transformation
        all_words: all words in the dictionary
        
        return nothing, just print top 5 positive and top 5 negative words for current tag
    """
    print('Tag:\t{}'.format(tag))
    
    coef = classifier.coef_[tags_classes.index(tag)]
    
    top_positive_words = [index_to_words[idx] for idx in coef.argsort()[-1:-10:-1]]# top-5 words sorted by the coefficiens.
    top_negative_words = [index_to_words[idx] for idx in coef.argsort()[:10]]# bottom-5 words  sorted by the coefficients.
    print('Top positive words:\t{}'.format(', '.join(top_positive_words)))
    print('Top negative words:\t{}\n'.format(', '.join(top_negative_words)))

___
## Read data:

train_df = pd.read_pickle(TRAIN_PATH)
val_df = pd.read_pickle(VAL_PATH)
test_df = pd.read_pickle(TEST_PATH)

# Load BOW features
with open(TRAIN_FEATURES_BOW, 'rb') as handle:
    X_train_BOW = pickle.load(handle)

with open(VAL_FEATURES_BOW, 'rb') as handle:
    X_val_BOW = pickle.load(handle)

with open(TEST_FEATURES_BOW, 'rb') as handle:
    X_test_BOW = pickle.load(handle)

# Load TF-IDF features
with open(TRAIN_FEATURES_TFIDF, 'rb') as handle:
    X_train_tfidf = pickle.load(handle)

with open(VAL_FEATURES_TFIDF, 'rb') as handle:
    X_val_tfidf = pickle.load(handle)

with open(TEST_FEATURES_TFIDF, 'rb') as handle:
    X_test_tfidf = pickle.load(handle)

# Load MultiLabelBinarizer target
with open(TRAIN_TARGET_EXPORT, 'rb') as handle:
    y_train = pickle.load(handle)

with open(VAL_TARGET_EXPORT, 'rb') as handle:
    y_val = pickle.load(handle)

# Load the MLB
with open('../data/objects/mlb.pkl', 'rb') as handle:
    mlb = pickle.load(handle)
    
# Load the BOW_vectorizer
with open('../data/objects/BOW_vectorizer.pkl', 'rb') as handle:
    BOW_vectorizer = pickle.load(handle)
    
# Load the tfidf_vectorizer
with open('../data/objects/tfidf_vectorizer.pkl', 'rb') as handle:
    tfidf_vectorizer = pickle.load(handle)

___
## Construct the base model:

lor = LogisticRegression(penalty='l2', C=1.0, solver='liblinear')

___
## Train model with Bag_Of_Words representaiton:

BOW_clf = make_pipeline(OneVsRestClassifier(lor))
BOW_clf.fit(X_train_BOW, y_train)

# Prepare data and metrics that would be used to evaluate the model
data_list = [[X_train_BOW, y_train, 'train'], [X_val_BOW, y_val, 'test']]
metrics_list = [accuracy_score, precision_score, recall_score, f1_score] 

# Use evaluate_model function to run evaluations
scores_BOW, predicts_BOW = evaluate_model(BOW_clf, data_list, metrics_list, "LogisticRegression", 'micro')
scores_BOW

# Use evaluate_model function to run evaluations
scores_BOW, predicts_BOW = evaluate_model(BOW_clf, data_list, metrics_list, "LogisticRegression", 'macro')
scores_BOW

# Use evaluate_model function to run evaluations
scores_BOW, predicts_BOW = evaluate_model(BOW_clf, data_list, metrics_list, "LogisticRegression", 'weighted')
scores_BOW

**Run some random samples to test the classifier** 

# Get the invers of the y_val predictions to find the predicted class
y_val_inv_BOW = mlb.inverse_transform(predicts_BOW[1].values) # (predicts) has the predictions of both X_train_BOW and X_val_BOW

samples = np.random.randint(1, val_df.shape[0], 5)
for i in samples:
    print("title: ", val_df['title'][i])
    print('True labels: ', val_df['tags'][i])
    print("Predicted labels: ", y_val_inv_BOW[i])
    print()

___
## Train model with TF-IDF representaiton:

tfidf_clf = make_pipeline(OneVsRestClassifier(lor))
tfidf_clf.fit(X_train_tfidf, y_train)

# Prepare data and metrics that would be used to evaluate the model
data_list = [[X_train_tfidf, y_train, 'train'], [X_val_tfidf, y_val, 'test']]
metrics_list = [accuracy_score, precision_score, recall_score, f1_score] 

# Use evaluate_model function to run evaluations
scores_tfidf, predicts_tfidf = evaluate_model(tfidf_clf, data_list, metrics_list, "LogisticRegression", 'micro')
scores_tfidf

# Use evaluate_model function to run evaluations
scores_tfidf, predicts_tfidf = evaluate_model(tfidf_clf, data_list, metrics_list, "LogisticRegression", 'macro')
scores_tfidf

# Use evaluate_model function to run evaluations
scores_tfidf, predicts_tfidf = evaluate_model(tfidf_clf, data_list, metrics_list, "LogisticRegression", 'weighted')
scores_tfidf

**Run some random samples to test the classifier** 

# Get the invers of the y_val predictions to find the predicted class
y_val_inv_tfidf = mlb.inverse_transform(predicts_tfidf[1].values) # (predicts) has the predictions of both X_train_BOW and X_val_BOW

samples = np.random.randint(1, val_df.shape[0], 5)
for i in samples:
    print("title: ", val_df['title'][i])
    print('True labels: ', val_df['tags'][i])
    print("Predicted labels: ", y_val_inv_tfidf[i])
    print()

___
## Analysis of most important features for each class:

**1. BOW classifier:**

BOW_reversed_vocab = {i:word for word, i in BOW_vectorizer.vocabulary_.items()}

print_words_for_tag(BOW_clf.named_steps['onevsrestclassifier'], 'c', mlb.classes, BOW_reversed_vocab)
print_words_for_tag(BOW_clf.named_steps['onevsrestclassifier'], 'c++', mlb.classes, BOW_reversed_vocab)
print_words_for_tag(BOW_clf.named_steps['onevsrestclassifier'], 'linux', mlb.classes, BOW_reversed_vocab)
print_words_for_tag(BOW_clf.named_steps['onevsrestclassifier'], 'javascript', mlb.classes, BOW_reversed_vocab)
print_words_for_tag(BOW_clf.named_steps['onevsrestclassifier'], 'python', mlb.classes, BOW_reversed_vocab)

**2. TF-IDF classifier:**

tfidf_reversed_vocab = {i:word for word, i in tfidf_vectorizer.vocabulary_.items()}

print_words_for_tag(tfidf_clf.named_steps['onevsrestclassifier'], 'c', mlb.classes, tfidf_reversed_vocab)
print_words_for_tag(tfidf_clf.named_steps['onevsrestclassifier'], 'c++', mlb.classes, tfidf_reversed_vocab)
print_words_for_tag(tfidf_clf.named_steps['onevsrestclassifier'], 'linux', mlb.classes, tfidf_reversed_vocab)
print_words_for_tag(tfidf_clf.named_steps['onevsrestclassifier'], 'javascript', mlb.classes, tfidf_reversed_vocab)
print_words_for_tag(tfidf_clf.named_steps['onevsrestclassifier'], 'python', mlb.classes, tfidf_reversed_vocab)

___
**As we can see; the classifier trained on the tfidf representation is more robust and accurate than the classifier trained on the BOW document representation; and so we can notice this clearly when we examine the most positive and negative words that affects our model. We can see that the words are somehow relevant to the tag or class it predicts in the tfidf classifier and so we can see the huge effect of this representation than the traditional BOW**
